Falta apenas o mergeSortSR!

+++ pilha.c
@@ -1,26 +1,28 @@
 #include "pilha.h"
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
+#include <stddef.h>
 
 void pilha_inicio (TPilha *p) {
     p->topo = -1;
 }
 
-int pilha_vazia (TPilha *p) {
+uint64_t pilha_vazia (TPilha *p) {
     if (p->topo == -1)
         return 1;
     else
         return 0;
 }
 
-int pilha_cheia (TPilha *p) {
+uint64_t pilha_cheia (TPilha *p) {
     if (p->topo == MAXTAM - 1)
         return 1;
     else
         return 0;
 }
 
-void empilhar (TPilha *p, int x) {
+void empilhar (TPilha *p, size_t x) {
     if (pilha_cheia(p)) {
         printf ("Pilha cheia!\n");
     }
@@ -30,8 +32,8 @@ void empilhar (TPilha *p, int x) {
     }
 }
 
-int desempilhar (TPilha *p) {
-    int aux;
+uint64_t desempilhar (TPilha *p) {
+    size_t aux;
     if (pilha_vazia(p)) {
         printf ("Pilha vazia!\n");
         return 0;
